Modernize the Windows setup.
This updates the Windows stuff based on the experience I've gained building
the downstream X.org libraries this far. The key new bit is that it looks like
we're really supposed to use $LIBRARY_PREFIX as our install prefix rather than
$PREFIX. I also run autoreconf which looks like it will be necessary for the
Windows builds. Not actually needed for this package since it just installs
macro files, but I want the build scripts to be as consistent as possible.